label_studio_ml/examples/yolov5/requirements.txt
# from mmDetection
boto3
click==7.1.2
label-studio-ml==1.0.8
opencv-contrib-python-headless
redis
rq==1.10.1
supervisor==4.2.2
# torch==1.12
# torchvision==0.13

# From YOLOv5
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics>=8.0.111

# uwsgi==2.0.19.1 use only on production
